+++ packages/system/src/index.ts
@@ -1,31 +1,31 @@
-import { AnimationProps } from "./animation";
-import { SpaceProps } from "./space";
-import { TypographyProps } from "./typography";
-import { LayoutProps } from "./layout";
-import { ColorProps } from "./color";
-import { FlexProps } from "./flex";
-import { BorderProps } from "./border";
-import { OutlineProps } from "./outline";
-import { PositionProps } from "./position";
-import { ShadowProps } from "./shadow";
-import { compose, StyledProps } from "./compose";
+import { AnimationProps } from "./props/animation";
+import { SpaceProps } from "./props/space";
+import { TypographyProps } from "./props/typography";
+import { LayoutProps } from "./props/layout";
+import { ColorProps } from "./props/color";
+import { FlexProps } from "./props/flex";
+import { BorderProps } from "./props/border";
+import { OutlineProps } from "./props/outline";
+import { PositionProps } from "./props/position";
+import { ShadowProps } from "./props/shadow";
+import { consumeStyleProps, StyledProps } from "./consumeStyleProps";
 import { ResponsiveStyle } from "./types";
 import { PseudoProps, normalizePseudo, isPseudoProps } from "./pseudo";
-import { GridProps } from "./grid";
-import { ListProps } from "./list";
-import { EffectProps } from "./effect";
-import { TextProps } from "./text";
-import { FontProps } from "./font";
-import { MaskProps } from "./mask";
-import { ColumnProps } from "./column";
-import { BackgroundProps } from "./background";
-import { FilterProps } from "./filter";
+import { GridProps } from "./props/grid";
+import { ListProps } from "./props/list";
+import { EffectProps } from "./props/effect";
+import { TextProps } from "./props/text";
+import { FontProps } from "./props/font";
+import { MaskProps } from "./props/mask";
+import { ColumnProps } from "./props/column";
+import { BackgroundProps } from "./props/background";
+import { FilterProps } from "./props/filter";
 
 export { StyledKeyType, isStyledProp } from "./keys";
 
 export * from "./types";
 
-export const all = compose();
+export const all = consumeStyleProps;
 
 export { normalizePseudo, isPseudoProps };
 export type {
