+++ packages/system/src/props/background.ts
@@ -1,13 +1,13 @@
-import { BackgroundKeys } from "./keys";
+import { BackgroundKeys } from "../keys";
 import {
   AddProperty,
   CSSProperties,
   CSSValue,
   ResponsiveStyle,
   ValueConverter,
-} from "./types";
-import { applyResponsiveStyles } from "./responsive";
-import { toCssUnit } from "./toCSS";
+} from "../types";
+import { applyResponsiveStyles } from "../responsive";
+import { toCssUnit } from "../toCSS";
 
 // eslint-disable-next-line @typescript-eslint/ban-types
 export type BackgroundProps<AutoPrefix extends string = string & {}> = Partial<
