+++ packages/system/src/props/font.ts
@@ -1,13 +1,13 @@
-import { toCssUnit } from "./toCSS";
-import { FontKeys } from "./keys";
+import { toCssUnit } from "../toCSS";
+import { FontKeys } from "../keys";
 import {
   ResponsiveStyle,
   CSSProperties,
   AddProperty,
   ThemeSystemType,
   ValueConverter,
-} from "./types";
-import { applyResponsiveStyles } from "./responsive";
+} from "../types";
+import { applyResponsiveStyles } from "../responsive";
 
 export type FontProps<T extends ThemeSystemType = ThemeSystemType> = Partial<
   CSSProperties<"font"> &
